Pin njh/ruby-mqtt 0.5.0
This commit changes the ruby-mqtt dependency to be pinned to the 0.5.0
version instead of a branched version. An exact pin was used to guard
against potential breaking changes in a newer version of the upstream
project (specifically around client connection handling). The previous
branched version included a change which avoids frequent hangs when
running under JRuby -- see https://github.com/njh/ruby-mqtt/pull/104.
By reverting to the upstream release, the JRuby hangs would be
reintroduced for now.