Adjust opendxl-broker 0.2.0+ Docker startup for Travis
Prior to the 0.2.0 release of opendxl-broker, the broker process ran as
the root user in the Docker container and the full set of files in the
keystore was present almost immediately after the 'docker run' command
to initiate the container had completed.

With the 0.2.0 release of the opendxl-broker, however, the broker
process runs as a custom user, dxl.  This makes the keystore files
inaccessible to the local Travis user via a Docker volume mount. Also,
with the 0.2.0 release, a slight delay may exist between the time the
'docker run' command completes and the keystore is fully populated.
These issues cause Ruby tests which depend upon connectivity to the
broker to fail.

With the changes in this commit, the Travis configuration eliminates
the use of a volume mount and instead now uses the 'docker cp' command
to copy keystore files out of the container before starting the Ruby
tests. This works around the file permission issue. Also, a step was
added to use netcat (via nc) in a loop to confirm that the broker port,
8883, is listening before the 'docker cp' command is run. This ensures
that the keystore is fully populated before any files are extracted from
it. These changes should allow Ruby integration tests to pass again.